refactor: group RPC clients with the domains that own them
lib/rpc mixed shared plumbing with per-domain clients, so a domain's code was
split across two folders. The ProfileService client now lives in lib/profile
and the IdentityService client in lib/auth, next to the types and query seams
they serve; lib/rpc keeps only the shared transport. This makes
lib/<domain>/{types,<domain>-rpc.server,<domain>-query} the recipe new
domains copy.

The MSW handlers also stop duplicating the auth service's test base URL and
Connect URL scheme — both now come from one shared module.
