+++ src/test/msw/identity.ts
@@ -6,6 +6,7 @@ import {
   ResolveUserResponseSchema,
   ResolveUserResponse_Resolution,
 } from '@fair-n-square-co/apis/fairnsquare/service/authx/v1alpha1/authx_api_pb'
+import { authServiceMethodUrl } from './auth-service'
 
 /**
  * MSW handlers for the auth service's `IdentityService`, mocked at the wire (§9)
@@ -17,11 +18,7 @@ import {
  * message shape wrong — which is precisely the contract worth guarding.
  */
 
-/** Matches `AUTH_SERVICE_BASE_URL` in the tests that install these handlers. */
-export const TEST_AUTH_SERVICE_BASE_URL = 'http://auth.test'
-
-// Connect addresses a unary method as POST <baseUrl>/<package>.<Service>/<Method>.
-const RESOLVE_USER_URL = `${TEST_AUTH_SERVICE_BASE_URL}/fairnsquare.service.authx.v1alpha1.IdentityService/ResolveUser`
+const RESOLVE_USER_URL = authServiceMethodUrl('IdentityService', 'ResolveUser')
 
 /** What the BFF actually put on the wire, so a test can assert on it. */
 export type RecordedResolveUser = Readonly<{
